move script paths to central config
