+++ ui/src/components/head/__inline__/.babelrc
