+++ ui/src/components/head/__inline__/loadApp.js
